+++ internal/service/commit/filter_shas_with_signatures.go
@@ -1,0 +1,66 @@
+package commit
+
+import (
+	pb "gitlab.com/gitlab-org/gitaly-proto/go"
+	"gitlab.com/gitlab-org/gitaly/internal/rubyserver"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/codes"
+)
+
+func (s *server) FilterShasWithSignatures(bidi pb.CommitService_FilterShasWithSignaturesServer) error {
+	firstRequest, err := bidi.Recv()
+	if err != nil {
+		return err
+	}
+
+	if err = verifyFirstFilterShasWithSignaturesRequest(firstRequest); err != nil {
+		return err
+	}
+
+	ctx := bidi.Context()
+	client, err := s.CommitServiceClient(ctx)
+	if err != nil {
+		return err
+	}
+
+	clientCtx, err := rubyserver.SetHeaders(ctx, firstRequest.GetRepository())
+	if err != nil {
+		return err
+	}
+
+	rubyBidi, err := client.FilterShasWithSignatures(clientCtx)
+	if err != nil {
+		return err
+	}
+
+	if err := rubyBidi.Send(firstRequest); err != nil {
+		return err
+	}
+
+	return rubyserver.ProxyBidi(
+		func() error {
+			request, err := bidi.Recv()
+			if err != nil {
+				return err
+			}
+
+			return rubyBidi.Send(request)
+		},
+		rubyBidi,
+		func() error {
+			response, err := rubyBidi.Recv()
+			if err != nil {
+				return err
+			}
+
+			return bidi.Send(response)
+		},
+	)
+}
+
+func verifyFirstFilterShasWithSignaturesRequest(in *pb.FilterShasWithSignaturesRequest) error {
+	if in.Repository == nil {
+		return grpc.Errorf(codes.InvalidArgument, "no repository given")
+	}
+	return nil
+}