+++ CHANGELOG.md
@@ -6,6 +6,8 @@ UNRELEASED
   https://gitlab.com/gitlab-org/gitaly/merge_requests/466
 - Fix a panic in ListFiles RPC when git process is killed abruptly
   https://gitlab.com/gitlab-org/gitaly/merge_requests/460
+- Implement CommitService::FilterShasWithSignatures
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/461
 
 v0.55.0